♻️ Extract all the data models into a single module.
